refactor tpm_core into two files for math and sync functions
